[DESK] Reimplement InstallScreenSaverW to behave similarly to Windows' one.
This function "just" changes the per-user SCRNSAVE.EXE registry value
to point to the new specified file, changes the SPI values and opens
the desk.cpl "Screensaver" property tab for letting the user change
the new selected screensaver properties.

In particular, it does *NOT* copy the specified file to System32 or
anything else (and doesn't verify that it is a valid PE executable).

Our previous implementation did none of that, and was also relying on
some private setupapi functions.

Now, we behave closer to Windows. However, we also verify that the
specified file actually exists, before changing the screensaver.